pyproject.toml
[project]
name = "upcast"
version = "0.2.3"
description = "AST tools"
authors = [{ name = "MrLYC", email = "github@m.mrlyc.com" }]
readme = "README.md"
requires-python = ">=3.9,<4.0"
license = { file = "LICENSE" }
dependencies = [
    "pydantic>=2.9.2",
    "click>=8.1.7",
    "astroid>=3.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.5.1",
    "pre-commit>=3.4.0",
    "tox>=4.11.1",
    "black>=24.10.0",
    "flake8>=7.1.1",
    "ipdb>=0.13.13",
    "ruff>=0.7.3",
]

[project.scripts]
upcast = "upcast.main:main"

[project.urls]
Repository = "https://github.com/mrlyc/upcast"
Documentation = "https://mrlyc.github.io/upcast/"

[tool.uv]
index-url = "https://mirrors.jlu.edu.cn/pypi/web/simple"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "prometheus-client>=0.20.0",
]

[tool.mypy]
files = ["upcast"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E",
    "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    "SIM115",
    "TRY003",
]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["upcast"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
